refactor: rename vitest-config → vite-utils
The lib now contains Vite plugins (rawSvgPlugin, svgTestResolverIfVitest),
utilities (monorepoTsconfigAliases), and Vitest config (baseVitestConfig).
'vite-utils' better describes its mixed purpose than 'vitest-config'.
