+++ apps/betterangels-admin/vite.config.mts
@@ -3,7 +3,7 @@ import tailwindcss from '@tailwindcss/postcss';
 import react from '@vitejs/plugin-react';
 import path from 'path';
 import { ProxyOptions, defineConfig } from 'vite';
-import { monorepoTsconfigAliases, rawSvgPlugin } from '../../libs/vitest-config/src/index';
+import { monorepoTsconfigAliases, rawSvgPlugin } from '../../libs/vite-utils/src/index';
 import {
   baseHrefPlugin,
   getBranchBasePath,
