+++ apps/storybook-react/.storybook/main.ts
@@ -3,7 +3,7 @@ import tailwindcss from '@tailwindcss/postcss';
 import { resolve } from 'path';
 import { mergeConfig, searchForWorkspaceRoot } from 'vite';
 import { getBranchBasePath } from '../../../tools/shared/get-base-path.mjs';
-import { monorepoTsconfigAliases } from '../../../libs/vitest-config/src/index';
+import { monorepoTsconfigAliases } from '../../../libs/vite-utils/src/index';
 import svgr from 'vite-plugin-svgr';
 import {
   PLATFORM_STORIES,
