+++ libs/ba-platform/web/vite.config.ts
@@ -1,5 +1,5 @@
 import { mergeConfig } from 'vitest/config';
-import { baseVitestConfig } from '../../../libs/vitest-config/src/index';
+import { baseVitestConfig } from '../../../libs/vite-utils/src/index';
 export default mergeConfig(baseVitestConfig, {
   resolve: { tsconfigPaths: true },
   test: { environment: 'jsdom' },
