+++ libs/react/betterangels-admin/vite.config.ts
@@ -1,3 +1,3 @@
 import { mergeConfig } from 'vitest/config';
-import { baseVitestConfig } from '../../../libs/vitest-config/src/index';
+import { baseVitestConfig } from '../../../libs/vite-utils/src/index';
 export default mergeConfig(baseVitestConfig, {});
