+++ libs/vite-utils/project.json
@@ -1,7 +1,7 @@
 {
-  "name": "vitest-config",
+  "name": "vite-utils",
   "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "libs/vitest-config/src",
+  "sourceRoot": "libs/vite-utils/src",
   "projectType": "library",
   "tags": ["type:util", "scope:shared"]
 }
