+++ libs/vite-utils/src/raw-svg-plugin.ts
