+++ libs/vite-utils/src/svg-test-resolver.ts
