+++ tsconfig.base.json
@@ -20,7 +20,7 @@
     "paths": {
       "@betterangels/*": ["./apps/betterangels/*"],
       "@monorepo/apollo": ["./libs/apollo/src/index.ts"],
-      "@monorepo/vitest-config": ["./libs/vitest-config/src/index.ts"],
+      "@monorepo/vite-utils": ["./libs/vite-utils/src/index.ts"],
       "@monorepo/fetch": ["./libs/fetch/src/index.ts"],
       "@monorepo/assets": ["./libs/assets/src/index.ts"],
       "@monorepo/ba-platform": ["./libs/ba-platform/src/index.ts"],
